H.Res. 380 Engrossed in House
Relating to consideration of the House amendment to the Senate amendment to the bill (H.R. 2642) to provide for the reform and continuation of agricultural and other programs of the Department of Agriculture through fiscal year 2018, and for other purposes, providing for consideration of the resolution (H. Res. 378) expressing the sense of the House of Representatives regarding certain provisions of the Senate amendment to H.R. 2642 relating to the Secretary of Agriculture's administration of tariff-rate quotas for raw and refined sugar, and providing for consideration of the resolution (H. Res. 379) expressing the sense of the House of Representatives regarding certain provisions of the Senate amendment to H.R. 2642 relating to crop insurance.

Version:  Engrossed in House (eh)
Date:     2013-10-11
Congress: 113

Source: https://www.govinfo.gov/content/pkg/BILLS-113hres380eh/xml/BILLS-113hres380eh.xml

Sponsored-By: F000450
Roll-Call: House 113-1-543 2013-10-11 Passed 219-193
Roll-Call: House 113-1-544 2013-10-11 Passed 223-189